Remove conflicting flag
The `--no-cache` and `--update` flags for `apk add` are mutually exclusive. Using `--no-cache` indicates that no index files are to be persisted locally from the remote repository; using `--update` means that the remote index file will be downloaded and persisted to `/var/cache/apk/`.

In the interests of keeping the final image size small, I imagine that the former is the desired behaviour.