use alpine 3.12 instead of 3.12.1
What do you think about using the latest `3.12` version of Alpine instead of pinning it? (the current version is `3.12.3`) Or do you prefer pinning it to the exact version? In that case: I could work on a PR to automatically query the latest exact Alpine version, is this something that would be useful?